Move custom validators below their field classes
